+++ src/static/modules/App.js
@@ -1,5 +1,4 @@
 
-
 import { MainContent } from './maincontent/MainContent.js';
 import { MainOverlay } from './mainoverlay/MainOverlay.js';
 // import { LeftPanel } from './leftpanel/LeftPanel.js';